requirements.txt
wheel
Appium-Python-Client==0.52
